+++ src/client/constants.ts
@@ -1,4 +1,4 @@
 export const CONTACT_PHONE_NUMBER = "+19178103314";
-export const FREEZE_DATE = new Date("2022-12-09T13:00:00-0500");
+export const LEADERBOARD_FREEZE_DATE = new Date("2022-12-09T13:00:00-0500");
 export const LEADERBOARD_REFRESH_INTERVAL = 300;
 export const LEADERBOARD_PAGE_SIZE = 20;
