+++ src/components/CountdownTimers.tsx
@@ -1,27 +1,10 @@
 import { useState, useEffect } from "react";
-import SimpleCountdownTimer from "./SimpleCountdown";
+import { SimpleCountdownTimer } from "./SimpleCountdown";
 import { CONTACT_PHONE_NUMBER } from "../client/constants";
 import { BillboardButton } from "./design-system";
 
-interface RecurringCountdownTimerProps {
-  intervalSeconds: number;
-}
-
 interface MainCountdownTimerProps {
   endDatetime: Date;
-}
-
-export function RecurringCountdownTimer({ intervalSeconds }: RecurringCountdownTimerProps) {
-  const INTERVAL = intervalSeconds * 1000;
-  const [targetTime, setTargetTime] = useState<Date>(new Date(Math.ceil(Date.now() / INTERVAL) * INTERVAL));
-
-  const onDone = () => {
-    // +1 to Date.now() to force us to the next interval in case we happen to call this function
-    // on the exact millisecond when the previous interval finishes.
-    setTargetTime(new Date(Math.ceil((Date.now() + 1) / INTERVAL) * INTERVAL));
-  };
-
-  return <SimpleCountdownTimer endDatetime={targetTime} onDoneWindowSeconds={2} onDone={onDone} format="minutesOnly" />;
 }
 
 export function MainCountdownTimer({ endDatetime }: MainCountdownTimerProps) {
